requirements.txt
google-generativeai>=0.7.0
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
langchain-text-splitters>=0.2.0
langgraph>=0.1.0
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
